requirements.txt
streamlit
pandas
openpyxl
Pillow
plotly
Babel